+++ core/src/main/scala/c4s/process/LoggerProcess.scala
@@ -9,9 +9,9 @@ import fs2.{text, Stream}
 
 private[process] final class LoggerProcess[F[_]: Sync](process: Process[F], logger: Logger[F]) extends Process[F] {
 
-  final def run(command: String, path: Option[Path]): F[ProcessResult[F]] =
+  override final def run(command: String, input: Option[Stream[F, Byte]], path: Option[Path]): F[ProcessResult[F]] =
     for {
-      result <- process.run(command, path)
+      result <- process.run(command, input, path)
       newResult <- logProcessResult(path, command, result)
     } yield newResult